Initial source code support for SDL3. No build system support.

+++ src/common/delay.h
@@ -24,7 +24,7 @@
 namespace love
 {
 
-void sleep(unsigned int ms);
+void sleep(double ms);
 
 } // namespace love
